Update README.md
This is a vehicle and driver management app that has been designed to make it easy for vehicle owners to manage their fleet of vehicles and drivers. With this app, vehicle owners can register drivers and assign them to specific vehicles, as well as monitor their activities and performance.The app provides a range of role-based functionalities, which empower drivers to report their work to the vehicle owner by changing their state, such as load, unload, departure, and arrival. This makes it easy for vehicle owners to track the progress of their drivers and ensure that they are meeting their targets.In addition to these features, the app also offers powerful tools like seat tracking, on-route maintenance, and parking management. These tools make it easy for vehicle owners to monitor the condition of their vehicles and ensure that they are being used efficiently.The app has been developed using Dart and Flutter, which are powerful tools for building high-quality mobile applications. It also integrates with a range of APIs, including Restful APIs, which ensures that it is easy to integrate with other systems and services.
Overall, this vehicle and driver management app is a powerful tool for vehicle owners who want to monitor and control their fleet more effectively. It is easy to use and provides a range of powerful features that cater to the specific requirements of vehicle owners.